+++ src/components/CategoriesContainer.test.jsx
